infra/pyproject.toml
[project]
name = "marky-chat-ui-infra"
version = "0.1.0"
description = "Infrastructure for Marky Chat UI"
dependencies = [
    "aws-cdk-lib",
    "constructs",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["marky_chat_ui_infra"] 
